+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased]
 ### Added
 ### Changed
+- accept `auth_token` in `Update` to allow obtaining releases from private GitHub repos
+- use GitHub api url instead of browser url to download assets so that auth can be used for private repos
+- accept headers in `Download` that can be used in GET request to download url (required for passing in auth token for private GitHub repos)
 ### Removed
 
 ## [0.6.0]